requirements.txt
pyyaml==6.0.1
torch==2.3.1
gymnasium==0.29.1
moviepy==1.0.3
pygame==2.5.2
matplotlib==3.8.2